Adição: `authentication.js` > `modules`
Dividi o `store` em um modulo de autenticação a fim de organizar as coisas.
Futuramente vou fazer o mesmo com as actions de books.
